requirements.txt
Django==3.2.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
pylint==2.9.5
pylint-django==2.4.4
mysqlclient==2.0.3